Placeholder fix for IE
IE does not support placeholders, so if the current value of the message box is equal to the current value of the placeholder, set the value to an empty string before updating the value of the placeholder.  If we don't do this, the placeholder plugin will assume that the user has entered text.