Make sure to rebind validation
Validation was not being called before ajax submit in some situations.  This was because the the validation plugin would not bind the onSubmit handler after the .validation() method was called a second time.  If a .unbind() was called on the form, the onSubmit validation handler would be removed, but not re-applied when .validation was being called.