Fix issue where unread message count was not showing.
Hide the unread message count by default.
Add/remove the unread_message class via js only when needed.
Give the element a class of 'unread_count' which is display: none by default.

It used to be that the UnreadMessages span was only being added by the template when there was an unread message.  However, the template is only refreshed when there is a conversation is added or removed.  Therefor, if there were no unread messages for a conversation when the template was loaded, the js was unable to find and update the UnreadMessages span for the conversation because it did not exist.