Fix a case where client login URLs were placed in transcript emails
This is a rare case where the client or operator didn't close the conversation and the CLI scripts closed it due to inactivity, and the client would have had to supplied their email address before starting the conversation.  The CLI scripts echo redirects for unit test reasons, which was placing the redirect URL in emails.    This change makes it so that only the "PHPT" environment will echo the redirect.

This change also short circuits the `requireClientLogin()` method in the CLI environment because that level of permission should be assumed in command line scripts.